requirements.txt
aiohttp==3.8.1
aiosignal==1.3.1
async-timeout==4.0.2
attrs==23.1.0
certifi==2023.5.7
charset-normalizer==2.1.1
frozenlist==1.3.3
giteapy==1.0.8
github==1.2.7
idna==3.4
mercurial==5.6.1
multidict==6.0.4
python-dateutil==2.8.2
python-dotenv==1.0.0
six==1.16.0
typing_extensions==4.6.3
urllib3==2.0.2
yarl==1.9.2
